refactor: city 패키지 생성
