+++ src/main/java/com/hubo/gillajabi/crawl/domain/service/WeatherCrawlServiceImpl.java
@@ -3,14 +3,14 @@ package com.hubo.gillajabi.crawl.domain.service;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hubo.gillajabi.crawl.domain.constant.ForecastType;
 import com.hubo.gillajabi.crawl.domain.constant.WeatherRedisConstants;
-import com.hubo.gillajabi.crawl.domain.entity.City;
+import com.hubo.gillajabi.city.domain.City;
 import com.hubo.gillajabi.crawl.infrastructure.dto.request.WeatherCurrentDto;
 import com.hubo.gillajabi.crawl.infrastructure.dto.request.WeatherTermDTO;
 import com.hubo.gillajabi.crawl.infrastructure.dto.response.AbstractApiResponse;
 import com.hubo.gillajabi.crawl.infrastructure.dto.response.ApiWeatherCurrentResponse;
 import com.hubo.gillajabi.crawl.infrastructure.dto.response.ApiWeatherMediumTermResponse;
 import com.hubo.gillajabi.crawl.infrastructure.exception.CrawlException;
-import com.hubo.gillajabi.crawl.infrastructure.persistence.CityRepository;
+import com.hubo.gillajabi.city.infrastructure.persistence.CityRepository;
 import com.hubo.gillajabi.crawl.infrastructure.util.helper.CrawlApiBuilderHelper;
 import com.hubo.gillajabi.global.common.dto.ApiProperties;
 import com.hubo.gillajabi.crawl.infrastructure.config.WeatherEndpointConfig;
