+++ src/main/java/com/hubo/gillajabi/crawl/domain/service/duru/RoadCityDuruService.java
@@ -1,8 +1,8 @@
 package com.hubo.gillajabi.crawl.domain.service.duru;
 
 import com.hubo.gillajabi.crawl.domain.constant.Province;
-import com.hubo.gillajabi.crawl.domain.entity.City;
-import com.hubo.gillajabi.crawl.infrastructure.persistence.CityRepository;
+import com.hubo.gillajabi.city.domain.City;
+import com.hubo.gillajabi.city.infrastructure.persistence.CityRepository;
 import com.hubo.gillajabi.crawl.infrastructure.dto.request.CityRequest;
 import com.hubo.gillajabi.crawl.infrastructure.dto.response.ApiCourseResponse;
 import com.hubo.gillajabi.crawl.infrastructure.exception.CrawlException;
